+++ build.gradle
@@ -47,7 +47,7 @@ application {
 }
 
 shadowJar {
-    archiveBaseName = "duke"
+    archiveBaseName = "energy"
     archiveClassifier = null
 }
