--- src/main/java/duke/controller/DialogBox.java
@@ -1,72 +1,0 @@
-package duke.controller;
-
-import java.io.IOException;
-import java.util.Collections;
-
-import javafx.collections.FXCollections;
-import javafx.collections.ObservableList;
-import javafx.fxml.FXML;
-import javafx.fxml.FXMLLoader;
-import javafx.geometry.Pos;
-import javafx.scene.Node;
-import javafx.scene.control.Label;
-import javafx.scene.image.Image;
-import javafx.scene.layout.HBox;
-import javafx.scene.paint.Color;
-import javafx.scene.paint.ImagePattern;
-import javafx.scene.shape.Circle;
-import javafx.scene.text.Font;
-import javafx.scene.text.FontWeight;
-import javafx.scene.text.TextAlignment;
-
-/**
- * An example of a custom control using FXML.
- * This control represents a dialog box consisting of an ImageView to represent the speaker's face and a label
- * containing text from the speaker.
- */
-public class DialogBox extends HBox {
-    @FXML
-    private Label dialog;
-    @FXML
-    private Circle displayPicture;
-
-    private DialogBox(String text, Image img) {
-        try {
-            FXMLLoader fxmlLoader =
-                    new FXMLLoader(MainWindow.class.getResource("/view/DialogBox.fxml"));
-            fxmlLoader.setController(this);
-            fxmlLoader.setRoot(this);
-            fxmlLoader.load();
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-        dialog.setText(text);
-        displayPicture.setFill(new ImagePattern(img));
-    }
-
-    /**
-     * Flips the dialog box such that the ImageView is on the left and text on the right.
-     */
-    private void flip() {
-        ObservableList<Node> tmp = FXCollections.observableArrayList(getChildren());
-        Collections.reverse(tmp);
-        getChildren().setAll(tmp);
-        dialog.setTextAlignment(TextAlignment.LEFT);
-        setAlignment(Pos.TOP_LEFT);
-    }
-
-    public static DialogBox getUserDialog(String text, Image img) {
-        return new DialogBox(text, img);
-    }
-
-    public static DialogBox getDukeDialog(String text, Image img, boolean hasException) {
-        var db = new DialogBox(text, img);
-        // Error messages should be coloured red and bolded
-        if (hasException) {
-            db.dialog.setTextFill(Color.RED);
-            db.dialog.setFont(Font.font(Font.getDefault().getName(), FontWeight.BOLD, 12));
-        }
-        db.flip();
-        return db;
-    }
-}
